requirements.txt
sqlalchemy~=2.0.32
requests~=2.32.3
python-dotenv~=1.0.1
flask~=3.0.3
pytest~=8.3.2